Refactor test for logout
